adding nodejs to the base package
